Rename: 인기 운동 루틴 panel 위치 이동
