+++ src/app/(with-container)/(exercise)/exercise/panel/VideoCard.tsx
