+++ src/app/(with-container)/(exercise)/exercise/panel/VideoInfo.tsx
