+++ src/app/(with-container)/(exercise)/exercise/panel/VideoInfoDetail.tsx
